Refactoring and fixed issue with dynamic fluid colors
