+++ src/main/java/appeng/items/parts/PartType.java
@@ -41,6 +41,9 @@ import appeng.core.AEConfig;
 import appeng.core.AppEng;
 import appeng.core.features.AEFeature;
 import appeng.core.localization.GuiText;
+import appeng.fluids.parts.PartFluidExportBus;
+import appeng.fluids.parts.PartFluidImportBus;
+import appeng.fluids.parts.PartFluidStorageBus;
 import appeng.integration.IntegrationRegistry;
 import appeng.integration.IntegrationType;
 import appeng.parts.automation.PartAnnihilationPlane;
@@ -49,13 +52,10 @@ import appeng.parts.automation.PartFormationPlane;
 import appeng.parts.automation.PartIdentityAnnihilationPlane;
 import appeng.parts.automation.PartImportBus;
 import appeng.parts.automation.PartLevelEmitter;
-import appeng.parts.automation.fluid.PartFluidExportBus;
-import appeng.parts.automation.fluid.PartFluidImportBus;
 import appeng.parts.misc.PartCableAnchor;
 import appeng.parts.misc.PartInterface;
 import appeng.parts.misc.PartInvertedToggleBus;
 import appeng.parts.misc.PartStorageBus;
-import appeng.parts.misc.PartFluidStorageBus;
 import appeng.parts.misc.PartToggleBus;
 import appeng.parts.networking.PartCableCovered;
 import appeng.parts.networking.PartCableGlass;
